requirements.txt
pandas
requests
boto3
apache-airflow-providers-amazon